documentation/requirements.txt
mutwo.ext-core
mutwo.ext-music
mutwo.ext-common-generators
mutwo.ext-mbrola
mutwo.ext-csound
mutwo.ext-reaper
mutwo.ext-midi
mutwo.ext-isis
mutwo.ext-abjad
mutwo.ext-ekmelily

sphinx==4.5.0
sphinx-autodoc-annotation==1.0-1
autodocsumm==0.2.4
insipid-sphinx-theme==0.2.8
autoclasstoc==1.5.1
